Move AbstractGenerateCommandOutput
